Add Redis Auth and mTLS support (#14)
This commit adds support for Redis Auth and mTLS to both `attache-check` and
`attache-control`. The number of configuration flags being added in this commit
necessitated a little bit of reorganization and cleanup to the config validation
code.

Similar to the aforementioned config cleanup, the large amount of new vars being
added to the `example/job-specification.hcl` necessitated the creation of an
`hcl` vars-file at `example/vars-file.hcl`. All new vars being added to the job
specification have been fully commented as well.

Fixes #7
Fixes #8
Part of #11